<#11> add support for parallel execution
Context:
During the processing of two files, Quick will group the their rows into groups before processing each group independently.

Proposal:
Because there is no dependencies between the groups processing, it will be fairly easy to run each one in parallel.
I do not expect to see a huge performance improvement, but the solution is easy to implement, and it will only be merged into the master branch if the increase in ram usage stays relatively the same.

results:
implemented using parallel collections, after some bench marks i noticed
that on average the execution time was reduced by only one 1s, which
is not that much, for now i do not count on merging this with bach in
the develop or the master branch.