<#10> checks for columns values
This Commit closes issue #10 in github.

After this commit a user can specify two new attributes
(checkColumnValueMatches and checkColumnValueExists) in the
ColumnDesription and ColumnIdentifier XML elements inside either
FixedOrderedRowDescription or DelimitedOrderedRowDescription.

checkColumnValueMatches is an optional attribute that takes a regex
and Quick will check whether the described column matches the regex
or no, if not the whole row will be considered as invalid and will
be displayed in the columns comparison table.

checkColumnValueExists is also an optional attribute that takes a
boolean true or false, with a default value of true. Quick will check
that the described column exists in the rows that match the row
description, if the column does not exist the whole row will be
considered as invalid and will be shown in the comparison table.

Quick will display ~ in the comparison table at the start of a
two columns comparison row and new to the column that failed the check.